Refactor interfaces and event implementation into interfaces.py
Fixes #9
