+++ src/zc/intid/subscribers.zcml
@@ -19,5 +19,5 @@
         for="zope.intid.interfaces.IIntIdEvent" />
     <subscriber
         handler=".subscribers.intIdEventNotify"
-        for="zc.intid.IIdEvent" />
+        for="zc.intid.interfaces.IIdEvent" />
 </configure>
